Update UIView-KIFAdditions.m to accomodate UIScrollView
The UIPickerView is a scroll View. It's also a UIPickerTableView
and that's causing the handler for scrolling all of the rows of
a UITableView into view so that KIF can account for all rows.
Problem with that is we're scrolling the UIPickerView and leaving
it in a state other than what it was last set. This address that
as a special case.